Don't include large externalServices.tsx file in initial bundle
