+++ web/src/site-admin/routes.tsx
@@ -2,16 +2,11 @@ import React from 'react'
 import { Redirect, RouteComponentProps } from 'react-router'
 import { lazyComponent } from '../util/lazyComponent'
 import { SiteAdminAreaRoute, SiteAdminAreaRouteContext } from './SiteAdminArea'
-import { codeHostExternalServices, nonCodeHostExternalServices } from '../components/externalServices/externalServices'
 import { Scalars } from '../graphql-operations'
 
 const ExternalServicesPage = lazyComponent(
     () => import('../components/externalServices/ExternalServicesPage'),
     'ExternalServicesPage'
-)
-const AddExternalServicesPage = lazyComponent(
-    () => import('../components/externalServices/AddExternalServicesPage'),
-    'AddExternalServicesPage'
 )
 const ExternalServicePage = lazyComponent(
     () => import('../components/externalServices/ExternalServicePage'),
@@ -53,15 +48,7 @@ export const siteAdminAreaRoutes: readonly SiteAdminAreaRoute[] = [
     },
     {
         path: '/external-services/new',
-        render: props => (
-            <AddExternalServicesPage
-                {...props}
-                routingPrefix="/site-admin"
-                afterCreateRoute="/site-admin/repositories?repositoriesUpdated"
-                codeHostExternalServices={codeHostExternalServices}
-                nonCodeHostExternalServices={nonCodeHostExternalServices}
-            />
-        ),
+        render: lazyComponent(() => import('./SiteAdminAddExternalServicesPage'), 'SiteAdminAddExternalServicesPage'),
         exact: true,
     },
     {
